+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 #### Visual Environment
 
+- [Nodes can be added to the graph by double-clicking the output ports of
+  existing nodes (or by clicking them with the right mouse button).][3346]
 - [Node Searcher preserves its zoom factor.][3327] The visible size of the node
   searcher and edited node is now fixed. It simplifies node editing on
   non-standard zoom levels.
@@ -116,6 +118,7 @@
 [3327]: https://github.com/enso-org/enso/pull/3327
 [3339]: https://github.com/enso-org/enso/pull/3339
 [3344]: https://github.com/enso-org/enso/pull/3344
+[3346]: https://github.com/enso-org/enso/pull/3346
 
 #### Enso Compiler